requirements.txt
streamlit==1.25.0
streamlit-webrtc==0.45.0
opencv-python-headless==4.7.0.72
tensorflow-cpu==2.10.0
numpy==1.23.5
av==10.0.0
mediapipe==0.9.0.1
